updating to use module system and have lib and bin crates
