+++ rls/debug/deps/server-1603fb14c8cdebf5.d
@@ -1,0 +1,7 @@
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/server-1603fb14c8cdebf5.rmeta: src/server.rs src/servo/mod.rs src/servo/http.rs
+
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/server-1603fb14c8cdebf5.d: src/server.rs src/servo/mod.rs src/servo/http.rs
+
+src/server.rs:
+src/servo/mod.rs:
+src/servo/http.rs:
