+++ rls/debug/deps/server-5d02369a58905871.d
@@ -1,0 +1,7 @@
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/server-5d02369a58905871.rmeta: src/server.rs src/http.rs src/servo.rs
+
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/server-5d02369a58905871.d: src/server.rs src/http.rs src/servo.rs
+
+src/server.rs:
+src/http.rs:
+src/servo.rs:
