+++ rls/debug/deps/server-af746b3b27e1c8e7.d
@@ -1,0 +1,7 @@
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/server-af746b3b27e1c8e7.rmeta: src/server.rs src/http.rs src/servo.rs
+
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/server-af746b3b27e1c8e7.d: src/server.rs src/http.rs src/servo.rs
+
+src/server.rs:
+src/http.rs:
+src/servo.rs:
