+++ rls/debug/deps/server-eba192fd131cd20b.d
@@ -1,0 +1,7 @@
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/server-eba192fd131cd20b.rmeta: src/server.rs src/servo/mod.rs src/servo/http.rs
+
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/server-eba192fd131cd20b.d: src/server.rs src/servo/mod.rs src/servo/http.rs
+
+src/server.rs:
+src/servo/mod.rs:
+src/servo/http.rs:
