+++ rls/debug/deps/servo-3dd22ba6dd8c2530.d
@@ -1,0 +1,6 @@
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/servo-3dd22ba6dd8c2530.rmeta: src/servo/mod.rs src/servo/http.rs
+
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/servo-3dd22ba6dd8c2530.d: src/servo/mod.rs src/servo/http.rs
+
+src/servo/mod.rs:
+src/servo/http.rs:
