+++ rls/debug/deps/servo-6142b9c20c9b25cd.d
@@ -1,0 +1,6 @@
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/servo-6142b9c20c9b25cd.rmeta: src/servo/mod.rs src/servo/http.rs
+
+/Users/karis/Programs/RustPrograms/simple_server/rls/debug/deps/servo-6142b9c20c9b25cd.d: src/servo/mod.rs src/servo/http.rs
+
+src/servo/mod.rs:
+src/servo/http.rs:
